Split Version and Process to App/Common
