+++ config/packages/twig.yaml
@@ -3,4 +3,4 @@ twig:
     form_themes:
         - 'bootstrap_3_layout.html.twig'
     globals:
-        version: '@App\Version\CacheVersionReader'
+        version: '@App\Common\Version\CacheVersionReader'
