+++ config/services_slave.yaml
@@ -1,0 +1,19 @@
+# This file is the entry point to configure your own services.
+# Files in the packages/ subdirectory configure your dependencies.
+
+# Put parameters here that don't need to change on each machine where the app is deployed
+# https://symfony.com/doc/current/best_practices/configuration.html#application-related-configuration
+parameters:
+    env(SLAVE_URL): "http://localhost"
+
+services:
+    # default configuration for services in *this* file
+    _defaults:
+        autowire: true      # Automatically injects dependencies in your services.
+        autoconfigure: true # Automatically registers your services as commands, event subscribers, etc.
+
+    App\Slave\:
+        resource: '../src/Slave/*'
+        exclude: '../src/{Client,Configuration.php,Device.php,Instruction.php,Probe.php}'
+
+    App\Slave\Client\FirepingClient: '@eight_points_guzzle.client.api_fireping'
