+++ src/Common/Process/ProcessInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Slave\Process;
+namespace App\Common\Process;
 
 interface ProcessInterface
 {
