+++ src/Common/Version/VersionReaderInterface.php
@@ -2,10 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Version;
-
-use Psr\Log\LoggerInterface;
-use Symfony\Component\Process\Process;
+namespace App\Common\Version;
 
 interface VersionReaderInterface
 {
