+++ src/Slave/Command/ProbeDispatcherCommand.php
@@ -4,15 +4,15 @@ declare(strict_types=1);
 
 namespace App\Slave\Command;
 
+use App\Common\Process\SymfonyProcessFactory;
+use App\Common\Version\GitVersionReader;
 use App\Slave\Configuration;
 use App\Slave\Instruction;
-use App\Slave\Process\SymfonyProcessFactory;
 use App\Slave\Task\FetchConfiguration;
 use App\Slave\Task\PublishStatistics;
 use App\Slave\Worker\Queue;
 use App\Slave\Worker\StatsManager;
 use App\Slave\Worker\WorkerManager;
-use App\Version\GitVersionReader;
 use Exception;
 use Psr\Log\LoggerInterface;
 use React\EventLoop\Factory;
